First two tests for buildEntityProto
